VAZRobin-JULOUNathanProject/requirements.txt
python-dotenv==1.0.1
requests==2.31.0
google-genai==0.4.1
Flask==3.0.2

Flask-SQLAlchemy>=3.1.1
PyMySQL>=1.1.0
cryptography>=42.0.0
Werkzeug>=3.0.0